FindDoxygen: build the generated file contents using temporary strings
Incrementally writing to a file, could result in write (permission
denied) errors.  Those errors are probably due to file-monitoring
processes having a temporary lock on a generated file.  The implemented
solution builds the content of the generated file as string and in the
end the built string is written to the generated file.

Fixes: #21190